Using terraform golang packages to build into the operator
