+++ deploy/crds/terraform.io_backends.yaml
@@ -1,0 +1,80 @@
+
+---
+apiVersion: apiextensions.k8s.io/v1beta1
+kind: CustomResourceDefinition
+metadata:
+  annotations:
+    controller-gen.kubebuilder.io/version: v0.2.2
+  creationTimestamp: null
+  name: backends.terraform.io
+spec:
+  additionalPrinterColumns:
+  - JSONPath: .status
+    description: Description of the current status
+    name: Status
+    type: string
+  group: terraform.io
+  names:
+    kind: Backend
+    listKind: BackendList
+    plural: backends
+    shortNames:
+    - bac
+    singular: backend
+  scope: Cluster
+  subresources:
+    status: {}
+  validation:
+    openAPIV3Schema:
+      description: Backend is the Schema for the Backends API
+      properties:
+        apiVersion:
+          description: 'APIVersion defines the versioned schema of this representation
+            of an object. Servers should convert recognized schemas to the latest
+            internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources'
+          type: string
+        kind:
+          description: 'Kind is a string value representing the REST resource this
+            object represents. Servers may infer this from the endpoint the client
+            submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds'
+          type: string
+        metadata:
+          type: object
+        spec:
+          description: BackendSpec defines the desired state of Backend
+          properties:
+            cacert_path:
+              description: EtcdV3 backend cacert path
+              type: string
+            cert_path:
+              description: EtcdV3 backend cert path
+              type: string
+            endpoints:
+              description: EtcdV3 backend endpoints
+              items:
+                type: string
+              type: array
+            key_path:
+              description: EtcdV3 backend key path
+              type: string
+            lock:
+              description: EtcdV3 backend lock
+              type: boolean
+            prefix:
+              description: EtcdV3 backend prefix
+              type: string
+          type: object
+        status:
+          type: string
+      type: object
+  version: v1alpha1
+  versions:
+  - name: v1alpha1
+    served: true
+    storage: true
+status:
+  acceptedNames:
+    kind: ""
+    plural: ""
+  conditions: []
+  storedVersions: []
