+++ deploy/crds/terraform.io_modules.yaml
@@ -1,0 +1,68 @@
+
+---
+apiVersion: apiextensions.k8s.io/v1beta1
+kind: CustomResourceDefinition
+metadata:
+  annotations:
+    controller-gen.kubebuilder.io/version: v0.2.2
+  creationTimestamp: null
+  name: modules.terraform.io
+spec:
+  additionalPrinterColumns:
+  - JSONPath: .status
+    description: Description of the current status
+    name: Status
+    type: string
+  group: terraform.io
+  names:
+    kind: Module
+    listKind: ModuleList
+    plural: modules
+    shortNames:
+    - mod
+    singular: module
+  scope: Namespaced
+  subresources:
+    status: {}
+  validation:
+    openAPIV3Schema:
+      description: Module is the Schema for the modules API
+      properties:
+        apiVersion:
+          description: 'APIVersion defines the versioned schema of this representation
+            of an object. Servers should convert recognized schemas to the latest
+            internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources'
+          type: string
+        kind:
+          description: 'Kind is a string value representing the REST resource this
+            object represents. Servers may infer this from the endpoint the client
+            submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds'
+          type: string
+        metadata:
+          type: object
+        spec:
+          description: ModuleSpec defines the desired state of Module
+          properties:
+            namespace_name:
+              description: Kubernetes namespace name
+              type: string
+            source:
+              description: Kubernetes namespace module source
+              enum:
+              - /var/lib/modules/kubernetes/namespace/
+              type: string
+          type: object
+        status:
+          type: string
+      type: object
+  version: v1alpha1
+  versions:
+  - name: v1alpha1
+    served: true
+    storage: true
+status:
+  acceptedNames:
+    kind: ""
+    plural: ""
+  conditions: []
+  storedVersions: []
