+++ pkg/terraform/terraform.go
@@ -2,94 +2,21 @@ package terraform
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/exec"
+
+	"github.com/mitchellh/cli"
+	"github.com/hashicorp/terraform/command"
+	backendInit "github.com/hashicorp/terraform/backend/init"
 )
-
-var TFPATH = os.Getenv("TFPATH")
-
-type Provider struct {
-	Provider map[string]interface{} `json:"provider"`
-}
-
-type Module struct {
-	Module map[string]interface{} `json:"module"`
-}
-
-type Terraform struct {
-	Terraform Backend `json:"terraform"`
-}
-
-type Backend struct {
-	Backend map[string]interface{} `json:"backend"`
-}
-
-// RenderProviderToTerraform takes an object, and attempts to construct the appropriate terraform json from it.
-func RenderProviderToTerraform(instance interface{}, providerName string) ([]byte, error) {
-	r := Provider{
-		Provider: map[string]interface{}{
-			providerName: instance,
-		},
-	}
-	b, err := json.MarshalIndent(r, "", "\t")
-	if err != nil {
-		return b, err
-	}
-	return b, nil
-}
-
-// RenderModuleToTerraform takes an object, and attempts to construct the appropriate terraform json from it.
-func RenderModuleToTerraform(instance interface{}, moduleName string) ([]byte, error) {
-	r := Module{
-		Module: map[string]interface{}{
-			moduleName: instance,
-		},
-	}
-	b, err := json.MarshalIndent(r, "", "\t")
-	if err != nil {
-		return b, err
-	}
-	return b, nil
-}
-
-// RenderModuleToTerraform takes an object, and attempts to construct the appropriate terraform json from it.
-func RenderBackendToTerraform(instance interface{}, backendName string) ([]byte, error) {
-	r := Terraform{
-		Backend{
-			Backend: map[string]interface{}{
-				backendName: instance,
-			},
-		},
-	}
-	b, err := json.MarshalIndent(r, "", "\t")
-	if err != nil {
-		return b, err
-	}
-	return b, nil
-}
-
-func WriteToFile(b []byte, namespace string, name string) error {
-	err := os.MkdirAll(TFPATH+"/"+namespace, os.ModePerm)
-	if err != nil {
-		return err
-	}
-
-	err = ioutil.WriteFile(TFPATH+"/"+namespace+"/"+name+".tf.json", b, 0755)
-	if err != nil {
-		return err
-	}
-	return nil
-}
 
 func TerraformNewWorkspace(namespace string) error {
 	var out bytes.Buffer
 	var stderr bytes.Buffer
 
 	cmd := exec.Command("terraform", "workspace", "new", namespace)
-	cmd.Dir = TFPATH + "/" + namespace
+	cmd.Dir = os.Getwd + "/" + namespace
 	cmd.Stdout = &out
 	cmd.Stderr = &stderr
 	cmd.Run()
@@ -103,7 +30,7 @@ func TerraformSelectWorkspace(namespace string) error {
 	var stderr bytes.Buffer
 
 	cmd := exec.Command("terraform", "workspace", "select", namespace)
-	cmd.Dir = TFPATH + "/" + namespace
+	cmd.Dir = os.Getwd + "/" + namespace
 	cmd.Stdout = &out
 	cmd.Stderr = &stderr
 	cmd.Run()
@@ -113,15 +40,21 @@ func TerraformSelectWorkspace(namespace string) error {
 }
 
 func TerraformInit(namespace string) error {
-	var out bytes.Buffer
-	var stderr bytes.Buffer
+	backendInit.Init(nil)
 
-	cmd := exec.Command("terraform", "init")
-	cmd.Dir = TFPATH + "/" + namespace
-	cmd.Stdout = &out
-	cmd.Stderr = &stderr
-	err := cmd.Run()
+  ui := &cli.BasicUi{
+		Reader:      os.Stdin,
+		Writer:      os.Stdout,
+		ErrorWriter: os.Stderr,
+  }
 
+  initCmd := &command.InitCommand{
+  	Meta: command.Meta{
+  		Ui: ui,
+  	},
+  }
+
+	exitStatus, err := initCmd.Run()
 	if err != nil {
 		fmt.Println(fmt.Sprint(err) + ": " + stderr.String())
 		return err
@@ -136,7 +69,7 @@ func TerraformValidate(namespace string) error {
 	var stderr bytes.Buffer
 
 	cmd := exec.Command("terraform", "validate")
-	cmd.Dir = TFPATH + "/" + namespace
+	cmd.Dir = os.Getwd + "/" + namespace
 	cmd.Stdout = &out
 	cmd.Stderr = &stderr
 	err := cmd.Run()
@@ -155,7 +88,7 @@ func TerraformPlan(namespace string) error {
 	var stderr bytes.Buffer
 
 	cmd := exec.Command("terraform", "plan")
-	cmd.Dir = TFPATH + "/" + namespace
+	cmd.Dir = os.Getwd + "/" + namespace
 	cmd.Stdout = &out
 	cmd.Stderr = &stderr
 	err := cmd.Run()
@@ -174,7 +107,7 @@ func TerraformApply(namespace string) error {
 	var stderr bytes.Buffer
 
 	cmd := exec.Command("terraform", "apply", "-auto-approve")
-	cmd.Dir = TFPATH + "/" + namespace
+	cmd.Dir = os.Getwd + "/" + namespace
 	cmd.Stdout = &out
 	cmd.Stderr = &stderr
 	err := cmd.Run()
